detail discount not showing fixed, tax disabled item detail page layout error fix download price bug fixed -change float to big decimal -change valueOf to parseDouble

+++ app/src/main/java/com/example/androidmobilestock/adapter/SettingHistoryItemAdapter.java
@@ -93,7 +93,7 @@ public class SettingHistoryItemAdapter extends RecyclerView.Adapter<SettingHisto
         myViewHolder.uom_textView.setText(item.UOM);
 
         // Assuming item.Price is a float
-        float price = item.Price;
+        double price = item.Price;
         String formattedPrice = String.format(Locale.US, "%.2f", price);
 
         //set item price
